+++ tencentcloud/vod/v20180717/models.go
@@ -7893,6 +7893,7 @@ type DescribeCDNStatDetailsRequestParams struct {
 	// <li>Bandwidth：带宽，单位为 Bps。</li>
 	// <li>Requests：请求数。</li>
 	// <li>QUICRequests：QUIC 请求数。</li>
+	// <li>IpVisit：IP 访问数。</li>
 	Metric *string `json:"Metric,omitempty" name:"Metric"`
 
 	// 起始时间，使用 [ISO 日期格式](https://cloud.tencent.com/document/product/266/11732#I)。
@@ -7983,6 +7984,7 @@ type DescribeCDNStatDetailsRequest struct {
 	// <li>Bandwidth：带宽，单位为 Bps。</li>
 	// <li>Requests：请求数。</li>
 	// <li>QUICRequests：QUIC 请求数。</li>
+	// <li>IpVisit：IP 访问数。</li>
 	Metric *string `json:"Metric,omitempty" name:"Metric"`
 
 	// 起始时间，使用 [ISO 日期格式](https://cloud.tencent.com/document/product/266/11732#I)。